requirements.txt
colorama==0.4.5
et-xmlfile==1.1.0
Flask==0.11.1
importlib-metadata==4.8.3
numpy
openpyxl==3.0.10
pandas
python-dateutil
Werkzeug
XlsxWriter
zipp
gunicorn
google-api-python-client
google-cloud-storage
google-cloud-bigquery
flask-bcrypt
Flask-Login
flask-sqlalchemy
Flask-Migrate